requirements.txt
requests
fake_useragent
names
customtkinter
pyyaml
